lab-04/aes_rsa_socket/requirements.txt
pycryptodome
Flask==2.3.2
python-socketio==5.9.0
python-engineio==4.7.1
flask-cors==4.0.0